change IMPORT_EXPORT_DIFF_BY_WORDS to IMPORT_EXPORT_DIFF_BY_CHARS
