+++ docs/installation.rst
@@ -70,10 +70,10 @@ You can use the following directives in your settings file:
     your own permission code. Default is ``None`` which means
     everybody can execute export action.
 
-``IMPORT_EXPORT_DIFF_BY_WORDS``
-    Global setting controls diff appearance. By setting it to ``True``,
+``IMPORT_EXPORT_DIFF_BY_CHARS``
+    Global setting controls diff appearance. By setting it to ``False``,
     diffs will be detected word by word instead of char by char.
-    Default is ``False``.
+    Default is ``True``.
 
 
 Example app
